TreeFloat, TreeInt 64 -> 32 bit values.

+++ src/lib.rs
@@ -17,8 +17,8 @@
 mod parsers;
 mod phylo;
 
-pub type TreeFloat = f64;
-pub type TreeInt = i64;
+pub type TreeFloat = f32;
+pub type TreeInt = i32;
 
 pub use parsers::TreeParseError;
 pub use parsers::newick::parse_newick;
